pyproject.toml
[tool.poetry]
name = "WorkKnow"
version = "0.1.0"
description = "Surface the Story of GitHub Actions Builds"
authors = ["Gregory M. Kapfhammer <gkapfham@allegheny.edu>"]
readme = "README.md"

[tool.poetry.scripts]
workknow = "workknow.main:cli"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.3.2"}
rich = "^10.5.0"
requests = "^2.25.1"
python-dotenv = "^0.18.0"
pandas = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
black = "^20.8b1"
pydocstyle = "^5.1.1"
flake8 = "^3.8.4"
taskipy = "^1.8.1"
pytest-cov = "^2.11.1"

[tool.taskipy.tasks]
black = { cmd = "black workknow tests --check", help = "Run the black checks for source code format" }
coverage = { cmd = "pytest -s --cov-config .coveragerc --cov-report term-missing --cov-report xml --cov --cov-branch", help = "Run coverage monitoring for the test suite" }
flake8 = { cmd = "flake8 workknow tests", help = "Run the flake8 checks for source code documentation" }
pydocstyle = { cmd = "pydocstyle workknow tests", help = "Run the pydocstyle checks for source code documentation" }
pylint = { cmd = "pylint workknow tests", help = "Run the pylint checks for source code documentation" }
test = { cmd = "pytest -x -s", help = "Run the pytest test suite (takes a long time)" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
